backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
rapidfuzz==3.5.2
scikit-learn==1.3.2
python-multipart==0.0.6
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
openai==1.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.9
phonenumbers==8.13.26
requests==2.31.0

